+++ apps/client/src/components/elements/ProfileCard.tsx
@@ -2,7 +2,7 @@ import { FC } from "react";
 
 import Image from "next/image";
 
-import { GetMyInfoQuery } from "@waifu-bot/graphql";
+import { GetMyInfoQuery } from "@waifu-bot/codegen";
 
 type ProfileCardProps = NonNullable<GetMyInfoQuery["me"]> & {
   profile: NonNullable<GetMyInfoQuery["me"]["profile"]>;