+++ apps/client/src/lib/contexts/authProvider.tsx
@@ -2,7 +2,7 @@ import { FC, PropsWithChildren, useState } from "react";
 
 import { useRouter } from "next/router";
 
-import { useGetMyInfoQuery } from "@waifu-bot/graphql";
+import { useGetMyInfoQuery } from "@waifu-bot/codegen";
 
 import { AuthContext, type IAuthContext } from "./authContext";